Lambda/telegram-lambda/requirements.txt
boto3>=1.26.0
telethon>=1.28.5
python-dotenv>=1.0.0 